tests/wpt/tests/html/semantics/embedded-content/the-iframe-element/iframe_javascript_url_remove_srcdoc.html

<!DOCTYPE html>
<title>javascript: URL in iframe src, removing srcdoc</title>
<link rel="help" href="https://html.spec.whatwg.org/multipage/#process-the-iframe-attributes">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<body>
<script>
setup({single_test: true});
let iframeLoaded = false;
window.javascriptUrlRan = false;
</script>
<iframe srcdoc="srcdoc" src="javascript:(() => { parent.javascriptUrlRan = true; })()" onload="iframeLoaded = true; this.onload = assert_unreached;"></iframe>
<script>
onload = () => {
  document.querySelector('iframe').removeAttribute('srcdoc');
  assert_true(iframeLoaded, "iframeLoaded");
  assert_false(javascriptUrlRan, "javascriptUrlRan");
  step_timeout(() => {
    assert_true(javascriptUrlRan, "javascriptUrlRan");
    done();
  }, 100); // Verify only one load event is fired on iframe
};
</script>
